requirements.txt
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.12.3
Django==4.2.17
django-bootstrap-v5==1.0.11
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
importlib_resources==6.5.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
packaging==24.2
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
soupsieve==2.6
sqlparse==0.5.3
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1